fix: fix tests for sdk
